+++ configuration_app.h
@@ -1,12 +1,12 @@
 #ifndef _CONFIGURATION_APP_H_
 #define _CONFIGURATION_APP_H_
 
-
+#include "Arduino.h"
 class Configuration {
     private:
-        uint16_t _configuration_pin;
+        uint8_t _configuration_pin;
     public:
-        Configuration(uint16_t _pin);
+        Configuration(uint8_t _pin);
         bool readConfiguration();
 };
